Large update, rebuilt code matrix generation and code mutation
Rebuilt the code matrix generation so that the a static large code matrix
is not used but the small code matrices are generated as needed.
Also changed code mutation so that it can generate mutant codes that only
have one gene/base changed.
Additionally, I added a new test model that currently runs, other models are
out of date as with all test.